Handle path parameter substitution on a per-component basis instead of over the whole string
When using `newPath.contains(_:)` and `newPath.replacingOccurences(of:with:)` on a String, you can run into issues when one key is a prefix of another key. For example, a path like `/:part/:partTwo` being substituted with `part: "valueOne", partTwo: "valueTwo"` will clobber the `partTwo` placeholder when substituting `part` (`/valueOne/valueOneTwo`). Relatedly, it also allows path parameters that only match a prefix of the placeholder in the path.

This PR first splits the path on `/` delimiters, only substitutes components that fully match the key, then re-joins the components into a path.